pyproject.toml
[tool.poetry]
name = "file_encryptor"
version = "0.1.17"
description = ""
authors = ["Michael Meier <michael_johannes.meier@daimler.com>"]
readme = "README.md"
packages = [{include = "src"}]
include = [
    { path = "config.yml"}
]

[tool.poetry.dependencies]
python = "^3.9"
cryptography= "^38.0.4"
python-dotenv= "^0.21.0"
loguru= "^0.6.0"
click = "^8.1.3"
pyyaml = "^6.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
file-encryptor = "src.cli:file_encryptor"
